Add Phylo.write() to output a tree.
