+++ src/newick.jl
@@ -25,7 +25,7 @@ equals = E"="
 
 @with_names begin
 @with_pre spc begin
-    str =  p"[a-zA-Z_][a-zA-Z\d_]*" | (quotes + p"[\"]+" + quotes)
+    str =  p"[a-zA-Z_][^\"\s\']*" | (quotes + p"[\"]+" + quotes)
     name = str | blank
     length_sep = E":" # |> "__length"
     length_val = Parse(p"-?(\d*\.?\d+|\d+\.\d*)([eE]-?\d+)?", Float64)
@@ -52,7 +52,7 @@ end
 @time open("test/Qian2016.tree", "r") do io
     parse_try(io, newick)
 end
-io = open("test/Qian2016.tree", "r")
+io = open("test/H1N1.newick", "r")
 out = read(io, String)
 parse_dbg(out, Trace(newick))
 @time parse_one(out, newick)
@@ -61,7 +61,167 @@ tree = "(A9CIT6:0.59280764,P51981:0.55926221,(Q8A861:0.99105703,((Q81IL5:0.76431
 tree = "(A9CIT6[&a=2]:0.59280764,P51981:0.55926221,(Q8A861:0.99105703,((Q81IL5:0.76431643,((((A1B198:0.94287572,(Q2U1E8:0.71410953,Q5LT20:0.55480466)0.975579:0.21734008)1.000000:0.55075633,(Q92YR6:1.11236546,(((Q13PB7:1.33807955,(Q161M1:1.17720944,A1AYL4:0.93440931)0.784619:0.18922325)0.878426:0.09769089,((((Q1ASJ4:1.28537785,(A8LS88:0.87558406,(C1DMY1:0.14671933,(P77215:0.02112667,Q8ZNF9:0.01593493)1.000000:0.35900384)1.000000:0.81055398)0.999947:0.27041496)0.403932:0.04809748)))))))))))));"
 parse_dbg(tree, Trace(newick))
 
+using Automa
+tokens = [
+    :lparens => re"\(",
+    :rparens => re"\)",
+    :comma => re",",
+    :name => re"[a-zA-Z_/][a-zA-Z0-9_/]*",
+    :quoted => re"\"[^\"]*\"",
+    :space => re"[ \t\n]*",
+    :metacomments => re"\[&.*\]"
+]
+@eval @enum Token error $(first.(tokens)...)
+make_tokenizer((error, 
+    [Token(i) => j for (i,j) in enumerate(last.(tokens))]
+)) |> eval
+
 =#
+
+abstract type NewickLike <: OutputType end
+struct Newick <: NewickLike end
+
+function outputnode!(io::IO, tree::AbstractTree{TT, RT, String}, node, ::Newick,
+                     ::Type{Nothing}) where {TT, RT}
+    print(io, "'", getnodename(tree, node), "'")
+    return nothing
+end
+
+function outputnode!(io::IO, tree::AbstractTree{TT, RT, <: Number}, node, ::Newick,
+                     ::Type{Nothing}) where {TT, RT}
+    print(io, getnodename(tree, node))
+    return nothing
+end
+
+function outputnode!(io::IO, tree::AbstractTree, node, ::Newick, ::Type{<: Dict})
+    print(io, "'", getnodename(tree, node), "'")
+    nd = getnodedata(tree, node)
+    if !isempty(nd)
+        print(io, "[&")
+        for (i, key) in enumerate(keys(nd))
+            if i > 1
+                print(io, ",")
+            end
+            value = nd[key]
+            if value isa String
+                print(io, key, "='", value, "'")
+            elseif value isa Number
+                print(io, key, "=", value)
+            elseif value isa Vector
+                print(io, key, "={")
+                for (j, elt) in enumerate(value)
+                    if j > 1
+                        print(io, ",")
+                    end
+                    if elt isa String
+                        print(io, "'", elt, "'")
+                    elseif elt isa Number
+                        print(io, elt)
+                    end
+                end
+                print(io, "}")
+            end
+        end
+        print(io, "]")
+    end
+    return nothing
+end
+
+function outputbranch!(io::IO, tree::AbstractTree, branch, ::NewickLike, ::Type{Nothing})
+    if haslength(tree, branch)
+        print(io, ":", getlength(tree, branch))
+    end
+    return nothing
+end
+
+function outputbranch!(io::IO, tree::AbstractTree, branch, ::NewickLike, ::Type{<: Dict})
+    if haslength(tree, branch)
+        print(io, ":")
+        bd = getbranchdata(tree, branch)
+        if !isempty(bd)
+            print(io, "[&")
+            for (i, key) in enumerate(keys(bd))
+                if i > 1
+                    print(io, ",")
+                end
+                value = nd[key]
+                if value isa String
+                    print(io, key, "='", value, "'")
+                elseif value isa Number
+                    print(io, key, "=", value)
+                elseif value isa Vector
+                    print(io, key, "={")
+                    for (j, elt) in enumerate(value)
+                        if j > 1
+                            print(io, ",")
+                        end
+                        if elt isa String
+                            print(io, "'", elt, "'")
+                        elseif elt isa Number
+                            print(io, elt)
+                        end
+                    end
+                    print(io, key, "}")
+                end
+            end
+            print(io, "]")
+        end
+        print(io, getlength(tree, branch))
+    end
+    return nothing
+end
+
+function outputsubtree!(io::IO, tree::T, node, ot::Newick) where T <: AbstractTree{OneTree, OneRoot}
+    if !isleaf(tree, node)
+        print(io, "(")
+        for (i, branch) in enumerate(getoutbounds(tree, node))
+            if i > 1
+                print(io, ",")
+            end
+            outputsubtree!(io, tree, dst(tree, branch), ot)
+            outputbranch!(io, tree, branch, ot, branchdatatype(T))
+        end
+        print(io, ")")
+    end
+    outputnode!(io, tree, node, ot, nodedatatype(T))
+    return nothing
+end
+
+function outputsubtree!(io::IO, tree::T, node, ot::Newick, exclude = []) where T <: AbstractTree{OneTree, Unrooted}
+    cs = getconnections(tree, node, exclude)
+    if !isempty(cs)
+        print(io, "(")
+        for (i, branch) in cs
+            if i > 1
+                print(io, ",")
+            end
+            outputsubtree!(io, tree, dst(tree, branch), ot, [branch])
+            outputbranch!(io, tree, branch, ot, branchdatatype(T))
+        end
+        print(io, ")")
+    end
+    outputnode!(io, tree, node, ot, nodedatatype(T))
+    return nothing
+end
+
+function outputtree!(io::IO, tree::AbstractTree{OneTree}, node, ot::Newick)
+    outputsubtree!(io::IO, tree::AbstractTree{OneTree}, node, ot::Newick)
+    print(io, ";")
+    return nothing
+end
+
+outputtree!(io::IO, tree::AbstractTree{OneTree, OneRoot}, ot::Newick) =
+    outputtree!(io, tree, getroot(tree), ot)
+
+write(io::IO, tree::T, ::Type{OT} = TT ≡ OneTree ? Newick : Nexus) where
+    {TT <: TreeType, T <: AbstractTree{TT}, OT <: NewickLike} =
+    outputtree!(io, tree, OT())
+
+function write(file::String, tree::AbstractTree)
+    open(file, "w") do io
+        write(io, tree)
+    end
+end
 
 using Tokenize
 using Tokenize.Lexers
@@ -580,7 +740,7 @@ function parsetrees(token, state, tokens, ::Type{TREE}, taxa) where
                                               t -> t.kind ∈ [T.LSQUARE, T.EQ])
         numTrees += 1
         if numTrees < 10
-            @info "Created a tree called \"$treename\""
+            @info "Created a tree called '$treename'"
         elseif numTrees == 10
             @info "[Stopping reporting on tree creation]"
         end
@@ -616,102 +776,3 @@ function parsetrees(token, state, tokens, ::Type{TREE}, taxa) where
     token, state = result
     return token, state, trees, treedata
 end
-
-function parsenexus(token, state, tokens, ::Type{TREE}) where
-    {RT, NL, N, B, TREE <: AbstractTree{OneTree, RT, NL, N, B}}
-    trees = missing
-    treedata = missing
-    taxa = Dict{NL, NL}()
-    while isBEGIN(token) || token.kind == T.LSQUARE
-        if token.kind == T.LSQUARE
-            while token.kind != T.RSQUARE
-                result = iterateskip(tokens, state)
-                isnothing(result) && return nothing
-                token, state = result
-            end
-            result = iterateskip(tokens, state)
-            isnothing(result) && return nothing
-            token, state = result
-        else
-            result = iterateskip(tokens, state)
-            isnothing(result) && return nothing
-            token, state = result
-            if checktosemi(isTAXA, token, state, tokens)
-                result = iterateskip(tokens, state)
-                isnothing(result) && return nothing
-                token, state = result
-                token, state = parsetaxa(token, state, tokens, taxa)
-            elseif checktosemi(isTREES, token, state, tokens)
-                result = iterateskip(tokens, state)
-                isnothing(result) && return nothing
-                token, state = result
-                token, state, trees, treedata = parsetrees(token, state, tokens, TREE, taxa)
-            else
-                @warn "Unexpected nexus block '$(untokenize(token))', skipping..."
-                result = iterateskip(tokens, state)
-                isnothing(result) && return nothing
-                token, state = result
-                while !checktosemi(isEND, token, state, tokens) && token.kind != T.ENDMARKER
-                    result = iterateskip(tokens, state)
-                    isnothing(result) && return nothing
-                    token, state = result
-                end
-                result = iterateskip(tokens, state)
-                isnothing(result) && return nothing
-                token, state = result
-            end
-        end
-    end
-
-    if token.kind != T.ENDMARKER
-        tokenerror(token, "end of file")
-    end
-    return TreeSet(trees, treedata)
-end
-
-function parsenexus(tokens::Tokenize.Lexers.Lexer,
-                    ::Type{TREE}) where {RT, NL, N, B,
-                                         TREE <: AbstractTree{OneTree, RT, NL, N, B}}
-    result = iterateskip(tokens)
-    isnothing(result) && return nothing
-    token, state = result
-    if token.kind == T.COMMENT && lowercase(untokenize(token)) == "#nexus"
-        result = iterateskip(tokens, state)
-        isnothing(result) && return nothing
-        token, state = result
-        return parsenexus(token, state, tokens, TREE)
-    else
-        error("Unexpected $(token.kind) token '$(untokenize(token))' " *
-              "at start of nexus file")
-    end
-end
-
-"""
-    parsenexus(io::IOBuffer, ::Type{TREE}) where TREE <: AbstractTree
-
-Parse an IOBuffer containing a nexus tree and convert into a phylogenetic
-tree of type TREE <: AbstractTree
-"""
-parsenexus(io::IOBuffer, ::Type{TREE}) where TREE <: AbstractTree =
-    parsenexus(tokenize(io), TREE)
-
-"""
-    parsenexus(io::IOStream, ::Type{TREE}) where TREE <: AbstractTree
-
-Parse an IOStream containing a nexus tree and convert into a phylogenetic
-tree of type TREE <: AbstractTree
-"""
-function parsenexus(ios::IOStream, ::Type{TREE}) where TREE <: AbstractTree
-    buf = IOBuffer()
-    print(buf, read(ios, String))
-    seek(buf, 0)
-    return parsenexus(buf, TREE)
-end
-
-"""
-    parsenexus(inp)
-
-Parse some input containing a nexus tree and convert into a phylogenetic
-tree of type RootedTree
-"""
-parsenexus(inp) = parsenexus(inp, RootedTree)
